requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pytest==7.4.0

# Kafka
confluent-kafka==2.1.1

# Data Processing
pandas==2.0.3
numpy==1.24.3
pyspark==3.4.1
pyspark-stubs==3.4.1.0

# Time Series Database
influxdb-client==1.36.1

# Utils
python-json-logger==2.0.7
schedule==1.2.0

# Visualization
grafana-api==1.0.7

# Data Generation
faker==19.3.0
